feat(mobile): convert notification screen to settings
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
